fix: foundry tests

+++ packages/core/test/foundry/data-verification-mechanism/fork-tests/FakeLifeCycle.ft.sol
@@ -22,7 +22,7 @@ contract FakeLifeCycle is CommonDataVerificationMechanismForkTest {
 
         // Mint fresh UMA and stake them.
         vm.prank(address(voting));
-        uint256 stakedNumOfTokens = gatMeetingNumOfTokens;
+        uint128 stakedNumOfTokens = gatMeetingNumOfTokens;
         votingToken.mint(TestAddress.account1, stakedNumOfTokens);
         vm.startPrank(TestAddress.account1);
         votingToken.approve(address(voting), stakedNumOfTokens);
@@ -60,8 +60,8 @@ contract FakeLifeCycle is CommonDataVerificationMechanismForkTest {
         assertEq(voting.getPrice(identifier, requestTime, ancillaryData), price);
 
         // Finally, considering we were the only voter, we should be able to work out the slashing amount precisely.
-        uint256 totalStakedAtVote = voting.cumulativeStake(); // Has not changed from when we staked.
-        uint256 slashPerTokenPerNoVote =
+        uint128 totalStakedAtVote = voting.cumulativeStake(); // Has not changed from when we staked.
+        uint128 slashPerTokenPerNoVote =
             voting.slashingLibrary().calcNoVoteSlashPerToken(
                 totalStakedAtVote,
                 stakedNumOfTokens,
